feat: tests updated to handle webhook
