+++ .env
@@ -9,7 +9,6 @@ CLIENT2_ADDRESS="0x976EA74026E726554dB657fA54763abd0C3a0aa9" # Account(6)
 PROPOSER_SIGNING_KEY="0x2a871d0798f97d79848a013d4936a73bf4cc922c825d33c1cf7073dff6d409c6" # key(9)
 PROPOSER_2_SIGNING_KEY="0x4bbbf85ce3377467afe5d46f804f221813b2bb87f24d81f60f1fcdbf7cbf4356" # key(7)
 DEPLOYER_SIGNING_KEY="0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80" # key(0)
-WEBHOOK_URL=
 AZURE_VAULT_URL=
 DEPLOYER_SIGNING_KEY_NAME=
 PROPOSER_SIGNING_KEY_NAME=
