+++ nightfall_test/Cargo.toml
@@ -38,5 +38,6 @@ figment = { version = "0.10.19", features = ["toml", "env"] }
 futures = "0.3.31"
 ethers-middleware = "2.0.14"
 uuid = { version = "1.16.0", features = ["v4"] }
+warp = "0.3.7"
